# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/projects":{"post":{"operationId":"ProjectsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectDto"}}}},"responses":{"201":{"description":""}},"tags":["Projects"]},"get":{"operationId":"ProjectsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Projects"]}},"/projects/{id}":{"get":{"operationId":"ProjectsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Projects"]},"patch":{"operationId":"ProjectsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectDto"}}}},"responses":{"200":{"description":""}},"tags":["Projects"]},"delete":{"operationId":"ProjectsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Projects"]}},"/blogs":{"post":{"operationId":"BlogsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBlogDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"BlogsController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/blogs/{id}":{"get":{"operationId":"BlogsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"BlogsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBlogDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"BlogsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/blog-comments":{"post":{"operationId":"BlogCommentsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBlogCommentDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"BlogCommentsController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/blog-comments/{id}":{"get":{"operationId":"BlogCommentsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"BlogCommentsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBlogCommentDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"BlogCommentsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/blog-subjects":{"post":{"operationId":"BlogSubjectsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBlogSubjectDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"BlogSubjectsController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/blog-subjects/{id}":{"get":{"operationId":"BlogSubjectsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"BlogSubjectsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBlogSubjectDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"BlogSubjectsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/replies":{"post":{"operationId":"RepliesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReplyDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"RepliesController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/replies/{id}":{"get":{"operationId":"RepliesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"RepliesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateReplyDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"RepliesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/visitors":{"post":{"operationId":"VisitorsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateVisitorDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"VisitorsController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/visitors/{id}":{"get":{"operationId":"VisitorsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"VisitorsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateVisitorDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"VisitorsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/illusts":{"post":{"operationId":"IllustsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateIllustDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"IllustsController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/illusts/{id}":{"get":{"operationId":"IllustsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"IllustsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateIllustDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"IllustsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"Wniko API","description":"","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateProjectDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"url":{"type":"string"},"headImageUrl":{"type":"string"},"githuburl":{"type":"string"},"starttime":{"type":"string"},"version":{"type":"string"},"technologies":{"type":"array","items":{"type":"string"}},"status":{"type":"number","description":"状态"},"isprivate":{"type":"boolean"}},"required":["name","description","url","headImageUrl","githuburl","starttime","version","technologies","status","isprivate"]},"UpdateProjectDto":{"type":"object","properties":{}},"CreateBlogDto":{"type":"object","properties":{}},"UpdateBlogDto":{"type":"object","properties":{}},"CreateBlogCommentDto":{"type":"object","properties":{}},"UpdateBlogCommentDto":{"type":"object","properties":{}},"CreateBlogSubjectDto":{"type":"object","properties":{}},"UpdateBlogSubjectDto":{"type":"object","properties":{}},"CreateReplyDto":{"type":"object","properties":{}},"UpdateReplyDto":{"type":"object","properties":{}},"CreateVisitorDto":{"type":"object","properties":{}},"UpdateVisitorDto":{"type":"object","properties":{}},"CreateIllustDto":{"type":"object","properties":{}},"UpdateIllustDto":{"type":"object","properties":{}}}}}